+++ src/main/java/com/modeul/web/service/StuffService.java
@@ -14,8 +14,8 @@ public interface StuffService {
 	List<StuffView> getViewAll();	// 전체 페이지 목록 조회용
 	List<StuffView> getViewAll(int page);	// 페이지 더보기용
 	List<StuffView> getViewAll(String query, int page);	// 검색용
-	List<StuffView> getViewAll(Integer categoryId, int page);	// 카테고리별 페이지 조회용
-	List<StuffView> getViewAll(String query, Integer categoryId, int page);	// 카테고리별 검색용 
+	List<StuffView> getViewAll(Long categoryId, int page);	// 카테고리별 페이지 조회용
+	List<StuffView> getViewAll(String query, Long categoryId, int page);	// 카테고리별 검색용 
 
 	Stuff getById(Long id);
